paratan2/pyproject.toml
[build-system]
requires = ["setuptools>=64", "setuptools_scm>=8", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paratan2"
version = "0.1.0"
description = "Engineering analysis tool for simple and tandem mirror fusion machines"
requires-python = ">=3.10"
dependencies = [
    "openmc",
    "numpy",
    "matplotlib",
    "pydantic>=2.0",
    "pyyaml",
]

[project.optional-dependencies]
gui = ["streamlit>=1.30"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools]
package-dir = {"" = "src"}
